+++ src/execution_engine/wrappers.rs
@@ -17,7 +17,7 @@ use llvm_sys::execution_engine::*;
 pub type LLVM_BSMMAllocateCodeSectionCallback =
     extern "C" fn(Opaque: *mut ::libc::c_void,
                   MM: LLVMMCJITMemoryManagerRef,
-                  Size: usize,
+                  Size: u32,
                   Alignment: ::libc::c_uint,
                   SectionID: ::libc::c_uint,
                   SectionName: *const ::libc::c_char) -> *mut u8;
@@ -26,7 +26,7 @@ pub type LLVM_BSMMAllocateCodeSectionCallback =
 pub type LLVM_BSMMAllocateDataSectionCallback =
     extern "C" fn(Opaque: *mut ::libc::c_void,
                   MM: LLVMMCJITMemoryManagerRef,
-                  Size: usize,
+                  Size: u32,
                   Alignment: ::libc::c_uint,
                   SectionID: ::libc::c_uint,
                   SectionName: *const ::libc::c_char,
@@ -68,14 +68,14 @@ extern "C" {
 
     pub fn LLVM_BSMMCallParentAllocateCodeSection(
         MM: LLVMMCJITMemoryManagerRef,
-        Size: usize,
+        Size: u32,
         Alignment: c_uint,
         SectionID: c_uint,
         SectionName: *const c_char) -> *mut u8;
 
     pub fn LLVM_BSMMCallParentAllocateDataSection(
         MM: LLVMMCJITMemoryManagerRef,
-        Size: usize,
+        Size: u32,
         Alignment: c_uint,
         SectionID: c_uint,
         SectionName: *const c_char,
